@@ -1,30 +1,28 @@
 name: flutter_quill_extensions
 description: Embed extensions for flutter_quill including image, video, formula and etc.
-version: 0.1.0
+version: 0.3.3
 homepage: https://bulletjournal.us/home/index.html
 repository: https://github.com/singerdmx/flutter-quill/tree/master/flutter_quill_extensions
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.0 <4.0.0"
   flutter: ">=3.0.0"
 
 dependencies:
   flutter:
     sdk: flutter
 
-  flutter_quill: ^6.0.0
+  flutter_quill: ^7.2.1
 
   image_picker: ^0.8.5+3
   photo_view: ^0.14.0
   video_player: ^2.4.2
-  youtube_player_flutter_quill: ^8.2.2
+  youtube_player_flutter: ^8.1.1
   gallery_saver: ^2.3.2
-  math_keyboard: ^0.1.6
-  string_validator: ^0.3.0
-
-#dependency_overrides:
-#  flutter_quill:
-#    path: ../
+  math_keyboard: ">=0.1.8 <0.3.0"
+  string_validator: ^1.0.0
+  universal_html: ^2.2.1
+  url_launcher: ^6.1.9
 
 dev_dependencies:
   flutter_test:
