@@ -1,32 +1,31 @@
 name: flutter_quill
 description: A rich text editor supporting mobile and web (Demo App @ bulletjournal.us)
-version: 6.1.6
+version: 7.2.11
 #author: bulletjournal
 homepage: https://bulletjournal.us/home/index.html
 repository: https://github.com/singerdmx/flutter-quill
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
-  flutter: ">=3.0.0"
+  sdk: ">=2.17.0 <4.0.0"
+  flutter: ">=3.10.0"
 
 dependencies:
   flutter:
     sdk: flutter
-  collection: ^1.16.0
+  collection: ^1.17.0
   flutter_colorpicker: ^1.0.3
-  flutter_keyboard_visibility: ^5.2.0
-  quiver: ^3.1.0
-  tuple: ^2.0.0
-  url_launcher: ^6.1.2
+  flutter_keyboard_visibility: ^5.4.0
+  quiver: ^3.2.1
+  url_launcher: ^6.1.9
   pedantic: ^1.11.1
-  characters: ^1.2.0
+  characters: ^1.2.1
   diff_match_patch: ^0.4.1
-  i18n_extension: ^6.0.0
-  device_info_plus: ^7.0.0
+  i18n_extension: ">=8.0.0 <10.0.0"
+  device_info_plus: ^9.0.0
   platform: ^3.1.0
   pasteboard: ^0.2.0
 
-dev_dependencies:
+  # Dependencies for testing utilities
   flutter_test:
     sdk: flutter
 
