@@ -1,10 +1,10 @@
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
-import 'package:tuple/tuple.dart';
 
 import '../models/documents/attribute.dart';
 import '../models/documents/nodes/block.dart';
 import '../models/documents/nodes/line.dart';
+import '../models/structs/vertical_spacing.dart';
 import '../utils/delta.dart';
 import 'box.dart';
 import 'controller.dart';
@@ -68,17 +68,19 @@ class EditableTextBlock extends StatelessWidget {
       required this.linkActionPicker,
       required this.cursorCont,
       required this.indentLevelCounts,
+      required this.clearIndents,
       required this.onCheckboxTap,
       required this.readOnly,
       this.onLaunchUrl,
       this.customStyleBuilder,
+      this.customLinkPrefixes = const <String>[],
       Key? key});
 
   final Block block;
   final QuillController controller;
   final TextDirection textDirection;
   final double scrollBottomInset;
-  final Tuple2 verticalSpacing;
+  final VerticalSpacing verticalSpacing;
   final TextSelection textSelection;
   final Color color;
   final DefaultStyles? styles;
@@ -91,8 +93,10 @@ class EditableTextBlock extends StatelessWidget {
   final CustomStyleBuilder? customStyleBuilder;
   final CursorCont cursorCont;
   final Map<int, int> indentLevelCounts;
+  final bool clearIndents;
   final Function(int, bool) onCheckboxTap;
   final bool readOnly;
+  final List<String> customLinkPrefixes;
 
   @override
   Widget build(BuildContext context) {
@@ -102,12 +106,12 @@ class EditableTextBlock extends StatelessWidget {
     return _EditableBlock(
         block: block,
         textDirection: textDirection,
-        padding: verticalSpacing as Tuple2<double, double>,
+        padding: verticalSpacing,
         scrollBottomInset: scrollBottomInset,
         decoration: _getDecorationForBlock(block, defaultStyles) ??
             const BoxDecoration(),
         contentPadding: contentPadding,
-        children: _buildChildren(context, indentLevelCounts));
+        children: _buildChildren(context, indentLevelCounts, clearIndents));
   }
 
   BoxDecoration? _getDecorationForBlock(
@@ -122,11 +126,14 @@ class EditableTextBlock extends StatelessWidget {
     return null;
   }
 
-  List<Widget> _buildChildren(
-      BuildContext context, Map<int, int> indentLevelCounts) {
+  List<Widget> _buildChildren(BuildContext context,
+      Map<int, int> indentLevelCounts, bool clearIndents) {
     final defaultStyles = QuillStyles.getStyles(context, false);
     final count = block.children.length;
     final children = <Widget>[];
+    if (clearIndents) {
+      indentLevelCounts.clear();
+    }
     var index = 0;
     for (final line in Iterable.castFrom<dynamic, Line>(block.children)) {
       index++;
@@ -143,15 +150,16 @@ class EditableTextBlock extends StatelessWidget {
             controller: controller,
             linkActionPicker: linkActionPicker,
             onLaunchUrl: onLaunchUrl,
+            customLinkPrefixes: customLinkPrefixes,
           ),
-          _getIndentWidth(),
+          _getIndentWidth(context),
           _getSpacingForLine(line, index, count, defaultStyles),
           textDirection,
           textSelection,
           color,
           enableInteractiveSelection,
           hasFocus,
-          MediaQuery.of(context).devicePixelRatio,
+          View.of(context).devicePixelRatio,
           cursorCont);
       final nodeTextDirection = getDirectionOfNode(line);
       children.add(Directionality(
@@ -162,45 +170,48 @@ class EditableTextBlock extends StatelessWidget {
 
   Widget? _buildLeading(BuildContext context, Line line, int index,
       Map<int, int> indentLevelCounts, int count) {
-    final defaultStyles = QuillStyles.getStyles(context, false);
+    final defaultStyles = QuillStyles.getStyles(context, false)!;
+    final fontSize = defaultStyles.paragraph?.style.fontSize ?? 16;
     final attrs = line.style.attributes;
+
     if (attrs[Attribute.list.key] == Attribute.ol) {
       return QuillNumberPoint(
         index: index,
         indentLevelCounts: indentLevelCounts,
         count: count,
-        style: defaultStyles!.leading!.style,
+        style: defaultStyles.leading!.style,
         attrs: attrs,
-        width: 32,
-        padding: 8,
+        width: fontSize * 2,
+        padding: fontSize / 2,
       );
     }
 
     if (attrs[Attribute.list.key] == Attribute.ul) {
       return QuillBulletPoint(
         style:
-            defaultStyles!.leading!.style.copyWith(fontWeight: FontWeight.bold),
-        width: 32,
+            defaultStyles.leading!.style.copyWith(fontWeight: FontWeight.bold),
+        width: fontSize * 2,
+        padding: fontSize / 2,
       );
     }
 
     if (attrs[Attribute.list.key] == Attribute.checked) {
       return CheckboxPoint(
-        size: 14,
+        size: fontSize,
         value: true,
         enabled: !readOnly,
         onChanged: (checked) => onCheckboxTap(line.documentOffset, checked),
-        uiBuilder: defaultStyles?.lists?.checkboxUIBuilder,
+        uiBuilder: defaultStyles.lists?.checkboxUIBuilder,
       );
     }
 
     if (attrs[Attribute.list.key] == Attribute.unchecked) {
       return CheckboxPoint(
-        size: 14,
+        size: fontSize,
         value: false,
         enabled: !readOnly,
         onChanged: (checked) => onCheckboxTap(line.documentOffset, checked),
-        uiBuilder: defaultStyles?.lists?.checkboxUIBuilder,
+        uiBuilder: defaultStyles.lists?.checkboxUIBuilder,
       );
     }
 
@@ -209,41 +220,43 @@ class EditableTextBlock extends StatelessWidget {
         index: index,
         indentLevelCounts: indentLevelCounts,
         count: count,
-        style: defaultStyles!.code!.style
+        style: defaultStyles.code!.style
             .copyWith(color: defaultStyles.code!.style.color!.withOpacity(0.4)),
-        width: 32,
+        width: fontSize * 2,
         attrs: attrs,
-        padding: 16,
+        padding: fontSize,
         withDot: false,
       );
     }
     return null;
   }
 
-  double _getIndentWidth() {
+  double _getIndentWidth(BuildContext context) {
+    final defaultStyles = QuillStyles.getStyles(context, false)!;
+    final fontSize = defaultStyles.paragraph?.style.fontSize ?? 16;
     final attrs = block.style.attributes;
 
     final indent = attrs[Attribute.indent.key];
     var extraIndent = 0.0;
     if (indent != null && indent.value != null) {
-      extraIndent = 16.0 * indent.value;
+      extraIndent = fontSize * indent.value;
     }
 
     if (attrs.containsKey(Attribute.blockQuote.key)) {
-      return 16.0 + extraIndent;
+      return fontSize + extraIndent;
     }
 
     var baseIndent = 0.0;
 
     if (attrs.containsKey(Attribute.list.key) ||
         attrs.containsKey(Attribute.codeBlock.key)) {
-      baseIndent = 32.0;
+      baseIndent = fontSize * 2;
     }
 
     return baseIndent + extraIndent;
   }
 
-  Tuple2 _getSpacingForLine(
+  VerticalSpacing _getSpacingForLine(
       Line node, int index, int count, DefaultStyles? defaultStyles) {
     var top = 0.0, bottom = 0.0;
 
@@ -252,22 +265,22 @@ class EditableTextBlock extends StatelessWidget {
       final level = attrs[Attribute.header.key]!.value;
       switch (level) {
         case 1:
-          top = defaultStyles!.h1!.verticalSpacing.item1;
-          bottom = defaultStyles.h1!.verticalSpacing.item2;
+          top = defaultStyles!.h1!.verticalSpacing.top;
+          bottom = defaultStyles.h1!.verticalSpacing.bottom;
           break;
         case 2:
-          top = defaultStyles!.h2!.verticalSpacing.item1;
-          bottom = defaultStyles.h2!.verticalSpacing.item2;
+          top = defaultStyles!.h2!.verticalSpacing.top;
+          bottom = defaultStyles.h2!.verticalSpacing.bottom;
           break;
         case 3:
-          top = defaultStyles!.h3!.verticalSpacing.item1;
-          bottom = defaultStyles.h3!.verticalSpacing.item2;
+          top = defaultStyles!.h3!.verticalSpacing.top;
+          bottom = defaultStyles.h3!.verticalSpacing.bottom;
           break;
         default:
           throw 'Invalid level $level';
       }
     } else {
-      late Tuple2 lineSpacing;
+      late VerticalSpacing lineSpacing;
       if (attrs.containsKey(Attribute.blockQuote.key)) {
         lineSpacing = defaultStyles!.quote!.lineSpacing;
       } else if (attrs.containsKey(Attribute.indent.key)) {
@@ -282,8 +295,8 @@ class EditableTextBlock extends StatelessWidget {
         // use paragraph linespacing as a default
         lineSpacing = defaultStyles!.paragraph!.lineSpacing;
       }
-      top = lineSpacing.item1;
-      bottom = lineSpacing.item2;
+      top = lineSpacing.top;
+      bottom = lineSpacing.bottom;
     }
 
     if (index == 1) {
@@ -294,7 +307,7 @@ class EditableTextBlock extends StatelessWidget {
       bottom = 0.0;
     }
 
-    return Tuple2(top, bottom);
+    return VerticalSpacing(top, bottom);
   }
 }
 
@@ -588,7 +601,7 @@ class RenderEditableTextBlock extends RenderEditableContainerBox
 }
 
 class _EditableBlock extends MultiChildRenderObjectWidget {
-  _EditableBlock(
+  const _EditableBlock(
       {required this.block,
       required this.textDirection,
       required this.padding,
@@ -601,13 +614,13 @@ class _EditableBlock extends MultiChildRenderObjectWidget {
 
   final Block block;
   final TextDirection textDirection;
-  final Tuple2<double, double> padding;
+  final VerticalSpacing padding;
   final double scrollBottomInset;
   final Decoration decoration;
   final EdgeInsets? contentPadding;
 
   EdgeInsets get _padding =>
-      EdgeInsets.only(top: padding.item1, bottom: padding.item2);
+      EdgeInsets.only(top: padding.top, bottom: padding.bottom);
 
   EdgeInsets get _contentPadding => contentPadding ?? EdgeInsets.zero;
 
