@@ -2,7 +2,7 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter_quill/flutter_quill.dart';
 import 'package:url_launcher/url_launcher.dart';
-import 'package:youtube_player_flutter_quill/youtube_player_flutter_quill.dart';
+import 'package:youtube_player_flutter/youtube_player_flutter.dart';
 
 class YoutubeVideoApp extends StatefulWidget {
   const YoutubeVideoApp(
