@@ -15,16 +15,22 @@ export 'embeds/toolbar/camera_button.dart';
 export 'embeds/toolbar/formula_button.dart';
 export 'embeds/toolbar/image_button.dart';
 export 'embeds/toolbar/image_video_utils.dart';
+export 'embeds/toolbar/media_button.dart';
 export 'embeds/toolbar/video_button.dart';
 export 'embeds/utils.dart';
 
 class FlutterQuillEmbeds {
-  static List<EmbedBuilder> builders(
-          {void Function(GlobalKey videoContainerKey)? onVideoInit}) =>
+  static List<EmbedBuilder> builders({
+    void Function(GlobalKey videoContainerKey)? onVideoInit,
+  }) =>
       [
         ImageEmbedBuilder(),
         VideoEmbedBuilder(onVideoInit: onVideoInit),
         FormulaEmbedBuilder(),
+      ];
+
+  static List<EmbedBuilder> webBuilders() => [
+        ImageEmbedBuilderWeb(),
       ];
 
   static List<EmbedButtonBuilder> buttons({
@@ -32,6 +38,10 @@ class FlutterQuillEmbeds {
     bool showVideoButton = true,
     bool showCameraButton = true,
     bool showFormulaButton = false,
+    String? imageButtonTooltip,
+    String? videoButtonTooltip,
+    String? cameraButtonTooltip,
+    String? formulaButtonTooltip,
     OnImagePickCallback? onImagePickCallback,
     OnVideoPickCallback? onVideoPickCallback,
     MediaPickSettingSelector? mediaPickSettingSelector,
@@ -39,54 +49,58 @@ class FlutterQuillEmbeds {
     FilePickImpl? filePickImpl,
     WebImagePickImpl? webImagePickImpl,
     WebVideoPickImpl? webVideoPickImpl,
-  }) {
-    return [
-      if (showImageButton)
-        (controller, toolbarIconSize, iconTheme, dialogTheme) => ImageButton(
-              icon: Icons.image,
-              iconSize: toolbarIconSize,
-              controller: controller,
-              onImagePickCallback: onImagePickCallback,
-              filePickImpl: filePickImpl,
-              webImagePickImpl: webImagePickImpl,
-              mediaPickSettingSelector: mediaPickSettingSelector,
-              iconTheme: iconTheme,
-              dialogTheme: dialogTheme,
-            ),
-      if (showVideoButton)
-        (controller, toolbarIconSize, iconTheme, dialogTheme) => VideoButton(
-              icon: Icons.movie_creation,
-              iconSize: toolbarIconSize,
-              controller: controller,
-              onVideoPickCallback: onVideoPickCallback,
-              filePickImpl: filePickImpl,
-              webVideoPickImpl: webImagePickImpl,
-              mediaPickSettingSelector: mediaPickSettingSelector,
-              iconTheme: iconTheme,
-              dialogTheme: dialogTheme,
-            ),
-      if ((onImagePickCallback != null || onVideoPickCallback != null) &&
-          showCameraButton)
-        (controller, toolbarIconSize, iconTheme, dialogTheme) => CameraButton(
-              icon: Icons.photo_camera,
-              iconSize: toolbarIconSize,
-              controller: controller,
-              onImagePickCallback: onImagePickCallback,
-              onVideoPickCallback: onVideoPickCallback,
-              filePickImpl: filePickImpl,
-              webImagePickImpl: webImagePickImpl,
-              webVideoPickImpl: webVideoPickImpl,
-              cameraPickSettingSelector: cameraPickSettingSelector,
-              iconTheme: iconTheme,
-            ),
-      if (showFormulaButton)
-        (controller, toolbarIconSize, iconTheme, dialogTheme) => FormulaButton(
-              icon: Icons.functions,
-              iconSize: toolbarIconSize,
-              controller: controller,
-              iconTheme: iconTheme,
-              dialogTheme: dialogTheme,
-            )
-    ];
-  }
+  }) =>
+      [
+        if (showImageButton)
+          (controller, toolbarIconSize, iconTheme, dialogTheme) => ImageButton(
+                icon: Icons.image,
+                iconSize: toolbarIconSize,
+                tooltip: imageButtonTooltip,
+                controller: controller,
+                onImagePickCallback: onImagePickCallback,
+                filePickImpl: filePickImpl,
+                webImagePickImpl: webImagePickImpl,
+                mediaPickSettingSelector: mediaPickSettingSelector,
+                iconTheme: iconTheme,
+                dialogTheme: dialogTheme,
+              ),
+        if (showVideoButton)
+          (controller, toolbarIconSize, iconTheme, dialogTheme) => VideoButton(
+                icon: Icons.movie_creation,
+                iconSize: toolbarIconSize,
+                tooltip: videoButtonTooltip,
+                controller: controller,
+                onVideoPickCallback: onVideoPickCallback,
+                filePickImpl: filePickImpl,
+                webVideoPickImpl: webImagePickImpl,
+                mediaPickSettingSelector: mediaPickSettingSelector,
+                iconTheme: iconTheme,
+                dialogTheme: dialogTheme,
+              ),
+        if ((onImagePickCallback != null || onVideoPickCallback != null) &&
+            showCameraButton)
+          (controller, toolbarIconSize, iconTheme, dialogTheme) => CameraButton(
+                icon: Icons.photo_camera,
+                iconSize: toolbarIconSize,
+                tooltip: cameraButtonTooltip,
+                controller: controller,
+                onImagePickCallback: onImagePickCallback,
+                onVideoPickCallback: onVideoPickCallback,
+                filePickImpl: filePickImpl,
+                webImagePickImpl: webImagePickImpl,
+                webVideoPickImpl: webVideoPickImpl,
+                cameraPickSettingSelector: cameraPickSettingSelector,
+                iconTheme: iconTheme,
+              ),
+        if (showFormulaButton)
+          (controller, toolbarIconSize, iconTheme, dialogTheme) =>
+              FormulaButton(
+                icon: Icons.functions,
+                iconSize: toolbarIconSize,
+                tooltip: formulaButtonTooltip,
+                controller: controller,
+                iconTheme: iconTheme,
+                dialogTheme: dialogTheme,
+              )
+      ];
 }
