@@ -1,9 +1,12 @@
 import 'dart:async';
 
-import 'package:tuple/tuple.dart';
-
+import '../../widgets/embeds.dart';
 import '../quill_delta.dart';
 import '../rules/rule.dart';
+import '../structs/doc_change.dart';
+import '../structs/history_changed.dart';
+import '../structs/offset_value.dart';
+import '../structs/segment_leaf_node.dart';
 import 'attribute.dart';
 import 'history.dart';
 import 'nodes/block.dart';
@@ -50,13 +53,12 @@ class Document {
     _rules.setCustomRules(customRules);
   }
 
-  final StreamController<Tuple3<Delta, Delta, ChangeSource>> _observer =
-      StreamController.broadcast();
+  final StreamController<DocChange> _observer = StreamController.broadcast();
 
   final History _history = History();
 
-  /// Stream of [Change]s applied to this document.
-  Stream<Tuple3<Delta, Delta, ChangeSource>> get changes => _observer.stream;
+  /// Stream of [DocChange]s applied to this document.
+  Stream<DocChange> get changes => _observer.stream;
 
   /// Inserts [data] in this document at specified [index].
   ///
@@ -158,7 +160,7 @@ class Document {
   }
 
   /// Returns all styles for each node within selection
-  List<Tuple2<int, Style>> collectAllIndividualStyles(int index, int len) {
+  List<OffsetValue<Style>> collectAllIndividualStyles(int index, int len) {
     final res = queryChild(index);
     return (res.node as Line).collectAllIndividualStyles(res.offset, len);
   }
@@ -167,6 +169,12 @@ class Document {
   List<Style> collectAllStyles(int index, int len) {
     final res = queryChild(index);
     return (res.node as Line).collectAllStyles(res.offset, len);
+  }
+
+  /// Returns all styles for any character within the specified text range.
+  List<OffsetValue<Style>> collectAllStylesWithOffset(int index, int len) {
+    final res = queryChild(index);
+    return (res.node as Line).collectAllStylesWithOffsets(res.offset, len);
   }
 
   /// Returns plain text within the specified text range.
@@ -216,19 +224,15 @@ class Document {
   }
 
   /// Given offset, find its leaf node in document
-  Tuple2<Line?, Leaf?> querySegmentLeafNode(int offset) {
+  SegmentLeafNode querySegmentLeafNode(int offset) {
     final result = queryChild(offset);
     if (result.node == null) {
-      return const Tuple2(null, null);
+      return const SegmentLeafNode(null, null);
     }
 
     final line = result.node as Line;
     final segmentResult = line.queryChild(result.offset, false);
-    if (segmentResult.node == null) {
-      return Tuple2(line, null);
-    }
-    final segment = segmentResult.node as Leaf;
-    return Tuple2(line, segment);
+    return SegmentLeafNode(line, segmentResult.node as Leaf?);
   }
 
   /// Composes [change] Delta into this document.
@@ -276,16 +280,16 @@ class Document {
     if (_delta != _root.toDelta()) {
       throw 'Compose failed';
     }
-    final change = Tuple3(originalDelta, delta, changeSource);
+    final change = DocChange(originalDelta, delta, changeSource);
     _observer.add(change);
     _history.handleDocChange(change);
   }
 
-  Tuple2 undo() {
+  HistoryChanged undo() {
     return _history.undo(this);
   }
 
-  Tuple2 redo() {
+  HistoryChanged redo() {
     return _history.redo(this);
   }
 
@@ -346,7 +350,13 @@ class Document {
   }
 
   /// Returns plain text representation of this document.
-  String toPlainText() => _root.children.map((e) => e.toPlainText()).join();
+  String toPlainText([
+    Iterable<EmbedBuilder>? embedBuilders,
+    EmbedBuilder? unknownEmbedBuilder,
+  ]) =>
+      _root.children
+          .map((e) => e.toPlainText(embedBuilders, unknownEmbedBuilder))
+          .join();
 
   void _loadDocument(Delta doc) {
     if (doc.isEmpty) {
