@@ -1,5 +1,6 @@
 import 'dart:math' as math;
 
+import '../../../widgets/embeds.dart';
 import '../../quill_delta.dart';
 import '../style.dart';
 import 'embeddable.dart';
@@ -224,7 +225,11 @@ class Text extends Leaf {
   String get value => _value as String;
 
   @override
-  String toPlainText() => value;
+  String toPlainText([
+    Iterable<EmbedBuilder>? embedBuilders,
+    EmbedBuilder? unknownEmbedBuilder,
+  ]) =>
+      value;
 }
 
 /// An embed node inside of a line in a Quill document.
@@ -257,7 +262,26 @@ class Embed extends Leaf {
   // Embed nodes are represented as unicode object replacement character in
   // plain text.
   @override
-  String toPlainText() => kObjectReplacementCharacter;
+  String toPlainText([
+    Iterable<EmbedBuilder>? embedBuilders,
+    EmbedBuilder? unknownEmbedBuilder,
+  ]) {
+    final builders = embedBuilders;
+
+    if (builders != null) {
+      for (final builder in builders) {
+        if (builder.key == value.type) {
+          return builder.toPlainText(this);
+        }
+      }
+    }
+
+    if (unknownEmbedBuilder != null) {
+      return unknownEmbedBuilder.toPlainText(this);
+    }
+
+    return Embed.kObjectReplacementCharacter;
+  }
 
   @override
   String toString() => '${super.toString()} ${value.type}';
