@@ -1,5 +1,6 @@
 import 'dart:collection';
 
+import '../../../widgets/embeds.dart';
 import '../../quill_delta.dart';
 import '../attribute.dart';
 import '../style.dart';
@@ -109,7 +110,10 @@ abstract class Node extends LinkedListEntry<Node> {
 
   Node newInstance();
 
-  String toPlainText();
+  String toPlainText([
+    Iterable<EmbedBuilder>? embedBuilders,
+    EmbedBuilder? unknownEmbedBuilder,
+  ]);
 
   Delta toDelta();
 
