@@ -35,6 +35,42 @@ extension Localization on String {
           'Next': 'Next',
           'Camera': 'Camera',
           'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
+          'Hex': 'Hex',
+          'Material': 'Material',
+          'Color': 'Color',
         },
         'en_us': {
           'Paste a link': 'Paste a link',
@@ -68,39 +104,113 @@ extension Localization on String {
           'Next': 'Next',
           'Camera': 'Camera',
           'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
+          'Hex': 'Hex',
+          'Material': 'Material',
+          'Color': 'Color',
         },
         'ar': {
           'Paste a link': 'نسخ الرابط',
           'Ok': 'نعم',
           'Select Color': 'اختار اللون',
-          'Gallery': 'الصور',
+          'Gallery': 'المعرض',
           'Link': 'الرابط',
           'Please first select some text to transform into a link.':
               'يرجى اختيار نص للتحويل إلى رابط',
           'Open': 'فتح',
-          'Copy': 'ينسخ',
+          'Copy': 'نسخ',
           'Remove': 'إزالة',
-          'Save': 'يحفظ',
+          'Save': 'حفظ',
           'Zoom': 'تكبير',
-          'Saved': 'أنقذ',
-          'Text': 'Text',
-          'What is entered is not a link': 'What is entered is not a link',
-          'Resize': 'Resize',
-          'Width': 'Width',
-          'Height': 'Height',
-          'Size': 'Size',
-          'Small': 'Small',
-          'Large': 'Large',
-          'Huge': 'Huge',
-          'Clear': 'Clear',
-          'Font': 'Font',
-          'Search': 'Search',
-          'matches': 'matches',
-          'showing match': 'showing match',
-          'Prev': 'Prev',
-          'Next': 'Next',
-          'Camera': 'Camera',
-          'Video': 'Video',
+          'Saved': 'تم الحفظ',
+          'Text': 'نص',
+          'What is entered is not a link': 'ما تم ادخاله ليس رابط',
+          'Resize': 'تحجيم',
+          'Width': 'عرض',
+          'Height': 'ارتفاع',
+          'Size': 'حجم',
+          'Small': 'صغير',
+          'Large': 'كبير',
+          'Huge': 'ضخم',
+          'Clear': 'تنظيف',
+          'Font': 'خط',
+          'Search': 'بحث',
+          'matches': 'تطابق',
+          'showing match': 'عرض التطابق',
+          'Prev': 'سابق',
+          'Next': 'تالي',
+          'Camera': 'كاميرا',
+          'Video': 'فيديو',
+          'Undo': 'تراجع',
+          'Redo': 'تقدم',
+          'Font family': 'عائلة الخط',
+          'Font size': 'حجم الخط',
+          'Bold': 'عريض',
+          'Subscript': 'نص سفلي',
+          'Superscript': 'نص علوي',
+          'Italic': 'مائل',
+          'Underline': 'تحته خط',
+          'Strike through': 'داخله خط',
+          'Inline code': 'كود بوسط السطر',
+          'Font color': 'لون الخط',
+          'Background color': 'لون الخلفية',
+          'Clear format': 'تنظيف التنسيق',
+          'Align left': 'محاذاة اليسار',
+          'Align center': 'محاذاة الوسط',
+          'Align right': 'محاذاة اليمين',
+          // i think it should be 'Justify with width'
+          // it is wrong in all properties
+          'Justify win width': 'Justify win width',
+          'Text direction': 'اتجاه النص',
+          'Header style': 'ستايل العنوان',
+          'Numbered list': 'قائمة مرقمة',
+          'Bullet list': 'قائمة منقطة',
+          'Checked list': 'قائمة للمهام',
+          'Code block': 'كود كامل',
+          'Quote': 'اقتباس',
+          'Increase indent': 'زيادة الهامش',
+          'Decrease indent': 'تنقيص الهامش',
+          'Insert URL': 'ادخل عنوان رابط',
+          'Visit link': 'زيارة الرابط',
+          'Enter link': 'ادخل رابط',
+          'Enter media': 'ادخل وسائط',
+          'Edit': 'تعديل',
+          'Apply': 'تطبيق',
+          'Hex': 'Hex',
+          'Material': 'Material',
+          'Color': 'اللون',
         },
         'da': {
           'Paste a link': 'Indsæt link',
@@ -134,6 +244,39 @@ extension Localization on String {
           'Next': 'Next',
           'Camera': 'Camera',
           'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
         },
         'de': {
           'Paste a link': 'Link hinzufügen',
@@ -168,6 +311,39 @@ extension Localization on String {
           'Next': 'Nächster',
           'Camera': 'Camera',
           'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
         },
         'fr': {
           'Paste a link': 'Coller un lien',
@@ -177,13 +353,13 @@ extension Localization on String {
           'Link': 'Lien',
           'Please first select some text to transform into a link.':
               "Veuillez d'abord sélectionner un texte à transformer en lien.",
-          'Open': 'Ouverte',
+          'Open': 'Ouvrir',
           'Copy': 'Copier',
           'Remove': 'Supprimer',
           'Save': 'Sauvegarder',
-          'Zoom': 'Zoom',
+          'Zoom': 'Zoomer',
           'Saved': 'Enregistrée',
-          'Text': 'Text',
+          'Text': 'Texte',
           'What is entered is not a link': "Ce qui est saisi n'est pas un lien",
           'Resize': 'Redimensionner',
           'Width': 'Largeur',
@@ -201,6 +377,39 @@ extension Localization on String {
           'Next': 'Suivant',
           'Camera': 'Caméra',
           'Video': 'Vidéo',
+          'Undo': 'Annuler',
+          'Redo': 'Refaire',
+          'Font family': 'Famille de police',
+          'Font size': 'Taille de police',
+          'Bold': 'Gras',
+          'Subscript': 'Indice',
+          'Superscript': 'Exposant',
+          'Italic': 'Italique',
+          'Underline': 'Souligné',
+          'Strike through': 'Barré',
+          'Inline code': 'Code en ligne',
+          'Font color': 'Couleur de police',
+          'Background color': 'Couleur de fond',
+          'Clear format': 'Effacer la mise en forme',
+          'Align left': 'Aligner à gauche',
+          'Align center': 'Aligner au centre',
+          'Align right': 'Aligner à droite',
+          'Justify win width': 'Justifier',
+          'Text direction': 'Direction du texte',
+          'Header style': "Style d'en-tête",
+          'Numbered list': 'Liste numérotée',
+          'Bullet list': 'Liste à puces',
+          'Checked list': 'Check-list',
+          'Code block': 'Bloc de code',
+          'Quote': 'Citation',
+          'Increase indent': 'Augmenter le retrait',
+          'Decrease indent': 'Diminuer le retrait',
+          'Insert URL': 'Insérer une URL',
+          'Visit link': 'Visiter',
+          'Enter link': 'Entrer un lien',
+          'Enter media': 'Entrer un média',
+          'Edit': 'Modifier',
+          'Apply': 'Appliquer',
         },
         'zh_cn': {
           'Paste a link': '粘贴链接',
@@ -209,7 +418,7 @@ extension Localization on String {
           'Gallery': '相簿',
           'Link': '链接',
           'Please first select some text to transform into a link.':
-              '请先选择一些要转化为链接的文本',
+              '请先选择需转化为链接的文本',
           'Open': '打开',
           'Copy': '复制',
           'Remove': '移除',
@@ -222,18 +431,183 @@ extension Localization on String {
           'Width': '宽度',
           'Height': '高度',
           'Size': '文字大小',
-          'Small': 'Small',
-          'Large': 'Large',
-          'Huge': 'Huge',
-          'Clear': 'Clear',
-          'Font': 'Font',
-          'Search': 'Search',
-          'matches': 'matches',
-          'showing match': 'showing match',
-          'Prev': 'Prev',
-          'Next': 'Next',
-          'Camera': 'Camera',
-          'Video': 'Video',
+          'Small': '小字号',
+          'Large': '大字号',
+          'Huge': '超大字号',
+          'Clear': '清除',
+          'Font': '字体',
+          'Search': '搜索',
+          'matches': '结果',
+          'showing match': '显示结果',
+          'Prev': '上一个',
+          'Next': '下一个',
+          'Camera': '拍照',
+          'Video': '录像',
+          'Undo': '撤销',
+          'Redo': '重做',
+          'Font family': '字体',
+          'Font size': '字号',
+          'Bold': '粗体',
+          'Subscript': '下标',
+          'Superscript': '上标',
+          'Italic': '斜体',
+          'Underline': '下划线',
+          'Strike through': '删除线',
+          'Inline code': '内联代码',
+          'Font color': '字体颜色',
+          'Background color': '背景颜色',
+          'Clear format': '清除格式',
+          'Align left': '左对齐',
+          'Align center': '居中对齐',
+          'Align right': '右对齐',
+          'Justify win width': '两端对齐',
+          'Text direction': '文本方向',
+          'Header style': '标题样式',
+          'Numbered list': '有序列表',
+          'Bullet list': '无序列表',
+          'Checked list': '任务列表',
+          'Code block': '代码块',
+          'Quote': '引言',
+          'Increase indent': '增加缩进',
+          'Decrease indent': '减少缩进',
+          'Insert URL': '插入链接',
+          'Visit link': '访问链接',
+          'Enter link': '输入链接',
+          'Enter media': '输入媒体',
+          'Edit': '编辑',
+          'Apply': '应用',
+        },
+        'zh_hk': {
+          'Paste a link': '貼上連結',
+          'Ok': '確定',
+          'Select Color': '選擇顏色',
+          'Gallery': '圖片庫',
+          'Link': '連結',
+          'Please first select some text to transform into a link.':
+              '請選擇文字以轉換為連結。',
+          'Open': '開啓',
+          'Copy': '複製',
+          'Remove': '移除',
+          'Save': '儲存',
+          'Zoom': '放大',
+          'Saved': '已儲存',
+          'Text': '文字',
+          'What is entered is not a link': '輸入的不是連結',
+          'Resize': '變更大小',
+          'Width': '寛',
+          'Height': '高',
+          'Size': '大小',
+          'Small': '小',
+          'Large': '大',
+          'Huge': '超大',
+          'Clear': '清除',
+          'Font': '字型',
+          'Search': '搜尋',
+          'matches': '符合',
+          'showing match': '顯示符合',
+          'Prev': '上一個',
+          'Next': '下一個',
+          'Camera': '相機',
+          'Video': '錄影',
+          'Undo': '撤銷',
+          'Redo': '重做',
+          'Font family': '字體',
+          'Font size': '字號',
+          'Bold': '粗體',
+          'Subscript': '下標',
+          'Superscript': '上標',
+          'Italic': '斜體',
+          'Underline': '下劃線',
+          'Strike through': '刪除線',
+          'Inline code': '內聯代碼',
+          'Font color': '字體顏色',
+          'Background color': '背景顏色',
+          'Clear format': '清除格式',
+          'Align left': '左對齊',
+          'Align center': '居中對齊',
+          'Align right': '右對齊',
+          'Justify win width': '兩端對齊',
+          'Text direction': '文本方向',
+          'Header style': '標題樣式',
+          'Numbered list': '有序列表',
+          'Bullet list': '無序列表',
+          'Checked list': '任務列表',
+          'Code block': '代碼塊',
+          'Quote': '引言',
+          'Increase indent': '增加縮進',
+          'Decrease indent': '減少縮進',
+          'Insert URL': '插入鏈接',
+          'Visit link': '訪問鏈接',
+          'Enter link': '輸入鏈接',
+          'Enter media': '輸入媒體',
+          'Edit': '編輯',
+          'Apply': '應用',
+        },
+        'ja': {
+          'Paste a link': 'リンクをペースト',
+          'Ok': '完了',
+          'Select Color': '色を選択',
+          'Gallery': '写真集',
+          'Link': 'リンク',
+          'Please first select some text to transform into a link.':
+              'まずリンクに変換する文字を選択してください.',
+          'Open': '開く',
+          'Copy': 'コピー',
+          'Remove': '削除',
+          'Save': '保存',
+          'Zoom': '拡大',
+          'Saved': '保存済み',
+          'Text': '文字',
+          'What is entered is not a link': '入力されたのはリンクではありません',
+          'Resize': 'サイズを調整',
+          'Width': '幅',
+          'Height': '高さ',
+          'Size': 'サイズ',
+          'Small': '小さい',
+          'Large': '大きい',
+          'Huge': 'でっかい',
+          'Clear': 'クリア',
+          'Font': 'フォント',
+          'Search': '検索',
+          'matches': '結果',
+          'showing match': '結果を表示',
+          'Prev': '前へ',
+          'Next': '次へ',
+          'Camera': 'カメラ',
+          'Video': 'ビデオ',
+          'Undo': '取り消し',
+          'Redo': 'やり直し',
+          'Font family': 'フォントファミリー',
+          'Font size': 'フォントサイズ',
+          'Bold': '太字',
+          'Subscript': '下付き',
+          'Superscript': '上付き',
+          'Italic': '斜体',
+          'Underline': '下線',
+          'Strike through': '取り消し線',
+          'Inline code': 'インラインコード',
+          'Font color': 'フォントカラー',
+          'Background color': 'ベースカラー',
+          'Clear format': 'クリアフォーマット',
+          'Align left': '左揃え',
+          'Align center': 'センターアライメント',
+          'Align right': '右揃え',
+          'Justify win width': '両端揃え',
+          'Text direction': '文字方向',
+          'Header style': 'タイトルスタイル',
+          'Numbered list': '順序付きリスト',
+          'Bullet list': '順序無しリスト',
+          'Checked list': 'チェックボックス',
+          'Code block': 'コード',
+          'Quote': '引用',
+          'Increase indent': 'インデントを増やす',
+          'Decrease indent': 'インデントを減らす',
+          'Insert URL': 'ハイパーリンクを挿入',
+          'Visit link': 'ハイパーリンクを訪問',
+          'Enter link': 'ハイパーリンクを輸入',
+          'Enter media': 'ミディアムを輸入',
+          'Edit': '編集',
+          'Apply': '応用',
         },
         'ko': {
           'Paste a link': '링크를 붙여넣어 주세요.',
@@ -251,21 +625,55 @@ extension Localization on String {
           'Saved': '저장되었습니다.',
           'Text': '텍스트',
           'What is entered is not a link': '입력한 내용은 링크가 아닙니다.',
-          'Resize': 'Resize',
-          'Width': 'Width',
-          'Height': 'Height',
-          'Small': 'Small',
-          'Large': 'Large',
-          'Huge': 'Huge',
-          'Clear': 'Clear',
-          'Font': 'Font',
-          'Search': 'Search',
-          'matches': 'matches',
-          'showing match': 'showing match',
-          'Prev': 'Prev',
-          'Next': 'Next',
-          'Camera': 'Camera',
-          'Video': 'Video',
+          'Resize': '크기조정',
+          'Width': '넓이',
+          'Height': '높이',
+          'Size': '크기',
+          'Small': '작게',
+          'Large': '크게',
+          'Huge': '매우크게',
+          'Clear': '초기화',
+          'Font': '글꼴',
+          'Search': '검색',
+          'matches': '결과',
+          'showing match': '결과 보기',
+          'Prev': '이전',
+          'Next': '다음',
+          'Camera': '카메라',
+          'Video': '비디오',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
         },
         'ru': {
           'Paste a link': 'Вставить ссылку',
@@ -299,6 +707,39 @@ extension Localization on String {
           'Next': 'Next',
           'Camera': 'Camera',
           'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
         },
         'es': {
           'Paste a link': 'Pega un enlace',
@@ -333,6 +774,39 @@ extension Localization on String {
           'Next': 'Siguiente',
           'Camera': 'Cámara',
           'Video': 'Vídeo',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
         },
         'tr': {
           'Paste a link': 'Bağlantıyı Yapıştır',
@@ -366,6 +840,39 @@ extension Localization on String {
           'Next': 'Devam',
           'Camera': 'Kamera',
           'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
         },
         'uk': {
           'Paste a link': 'Вставити посилання',
@@ -399,6 +906,39 @@ extension Localization on String {
           'Next': 'Next',
           'Camera': 'Camera',
           'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
         },
         'pt': {
           'Paste a link': 'Colar um link',
@@ -433,6 +973,39 @@ extension Localization on String {
           'Next': 'Next',
           'Camera': 'Camera',
           'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
         },
         'pt_br': {
           'Paste a link': 'Colar um link',
@@ -461,12 +1034,45 @@ extension Localization on String {
           'Clear': 'Limpar',
           'Font': 'Fonte',
           'Search': 'Buscar',
-          'matches': 'matches',
-          'showing match': 'showing match',
+          'matches': 'resultado(s)',
+          'showing match': 'mostrando resultado',
           'Prev': 'Anterior',
           'Next': 'Próximo',
-          'Camera': 'Camera',
+          'Camera': 'Câmera',
           'Video': 'Vídeo',
+          'Undo': 'Desfazer',
+          'Redo': 'Refazer',
+          'Font family': 'Fonte',
+          'Font size': 'Tamanho da fonte',
+          'Bold': 'Negrito',
+          'Subscript': 'Subscrito',
+          'Superscript': 'Sobrescrito',
+          'Italic': 'Itálico',
+          'Underline': 'Sublinhado',
+          'Strike through': 'Tachado',
+          'Inline code': 'Inline code',
+          'Font color': 'Cor da fonte',
+          'Background color': 'Cor do fundo',
+          'Clear format': 'Limpar formatação',
+          'Align left': 'Texto à esquerda',
+          'Align center': 'Centralizar',
+          'Align right': 'Texto à direita',
+          'Justify win width': 'Justificado',
+          'Text direction': 'Direção do texto',
+          'Header style': 'Estilo de cabeçalho',
+          'Numbered list': 'Numeração',
+          'Bullet list': 'Marcadores',
+          'Checked list': 'Lista de verificação',
+          'Code block': 'Code block',
+          'Quote': 'Citação',
+          'Increase indent': 'Aumentar recuo',
+          'Decrease indent': 'Diminuir recuo',
+          'Insert URL': 'Inserir URL',
+          'Visit link': 'Visitar link',
+          'Enter link': 'Inserir link',
+          'Enter media': 'Inserir mídia',
+          'Edit': 'Editar',
+          'Apply': 'Aplicar',
         },
         'pl': {
           'Paste a link': 'Wklej link',
@@ -501,6 +1107,39 @@ extension Localization on String {
           'Next': 'Next',
           'Camera': 'Camera',
           'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
         },
         'vi': {
           'Paste a link': 'Chèn liên kết',
@@ -535,6 +1174,39 @@ extension Localization on String {
           'Next': 'Next',
           'Camera': 'Camera',
           'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
         },
         'ur': {
           'Paste a link': 'لنک پیسٹ کریں',
@@ -568,6 +1240,39 @@ extension Localization on String {
           'Next': 'Next',
           'Camera': 'Camera',
           'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
         },
         'id': {
           'Paste a link': 'Tempel tautan',
@@ -601,6 +1306,39 @@ extension Localization on String {
           'Next': 'Next',
           'Camera': 'Camera',
           'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
         },
         'no': {
           'Paste a link': 'Lim inn lenke',
@@ -634,6 +1372,39 @@ extension Localization on String {
           'Next': 'Next',
           'Camera': 'Camera',
           'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
         },
         'fa': {
           'Paste a link': 'جایگذاری لینک',
@@ -667,6 +1438,39 @@ extension Localization on String {
           'Next': 'بعدی',
           'Camera': 'دوربین',
           'Video': 'ویدیو',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
         },
         'hi': {
           'Paste a link': 'लिंक पेस्ट करें',
@@ -700,6 +1504,39 @@ extension Localization on String {
           'Next': 'Next',
           'Camera': 'Camera',
           'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
         },
         'nl': {
           'Paste a link': 'Plak een link',
@@ -733,39 +1570,39 @@ extension Localization on String {
           'Next': 'Next',
           'Camera': 'Camera',
           'Video': 'Video',
-        },
-        'zh_hk': {
-          'Paste a link': '貼上連結',
-          'Ok': '確定',
-          'Select Color': '選擇顏色',
-          'Gallery': '圖片庫',
-          'Link': '連結',
-          'Please first select some text to transform into a link.':
-              '請選擇文字以轉換為連結。',
-          'Open': '開啓',
-          'Copy': '複製',
-          'Remove': '移除',
-          'Save': '儲存',
-          'Zoom': '放大',
-          'Saved': '已儲存',
-          'Text': '文字',
-          'What is entered is not a link': '輸入的不是連結',
-          'Resize': '變更大小',
-          'Width': '寛',
-          'Height': '高',
-          'Size': '大小',
-          'Small': '小',
-          'Large': '大',
-          'Huge': '超大',
-          'Clear': '清除',
-          'Font': '字型',
-          'Search': '搜尋',
-          'matches': '符合',
-          'showing match': '顯示符合',
-          'Prev': '上一個',
-          'Next': '下一個',
-          'Camera': '相機',
-          'Video': '錄影',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
         },
         'sr': {
           'Paste a link': 'Nalepi vezu',
@@ -799,6 +1636,375 @@ extension Localization on String {
           'Next': 'Next',
           'Camera': 'Camera',
           'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
+        },
+        'cs': {
+          'Paste a link': 'Vložte odkaz',
+          'Ok': 'Ok',
+          'Select Color': 'Vyberte barvu',
+          'Gallery': 'Galerie',
+          'Link': 'Odkaz',
+          'Please first select some text to transform into a link.':
+              'Nejprve vyberte nějaký text, který chcete převést na odkaz.',
+          'Open': 'Otevřít',
+          'Copy': 'Kopírovat',
+          'Remove': 'Odstranit',
+          'Save': 'Uložit',
+          'Zoom': 'Přiblížit',
+          'Saved': 'Uloženo',
+          'Text': 'Text',
+          'What is entered is not a link': 'Zadaný vstup není odkaz',
+          'Resize': 'Změnit velikost',
+          'Width': 'Šířka',
+          'Height': 'Výška',
+          'Size': 'Velikost',
+          'Small': 'Malý',
+          'Large': 'Velký',
+          'Huge': 'Obrovský',
+          'Clear': 'Smazat',
+          'Font': 'Písmo',
+          'Search': 'Hledat',
+          'matches': 'odpovídá',
+          'showing match': 'zobrazuje odpovídající',
+          'Prev': 'Předchozí',
+          'Next': 'Další',
+          'Camera': 'Kamera',
+          'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
+        },
+        'he': {
+          'Paste a link': 'הדבק את הלינק',
+          'Ok': 'אוקי',
+          'Select Color': 'בחר צבע',
+          'Gallery': 'גלריה',
+          'Link': 'לינק',
+          'Please first select some text to transform into a link.':
+              'בבקשה תחילה בחר טקסט להפיכה ללינק',
+          'Open': 'פתח',
+          'Copy': 'העתק',
+          'Remove': 'מחק',
+          'Save': 'שמור',
+          'Zoom': 'זום',
+          'Saved': 'נשמר',
+          'Text': 'טקסט',
+          'What is entered is not a link': 'מה שהוכנס הוא לא לינק',
+          'Resize': 'שנה גודל',
+          'Width': 'רוחב',
+          'Height': 'גובה',
+          'Size': 'גודל',
+          'Small': 'קטן',
+          'Large': 'גדול',
+          'Huge': 'ענק',
+          'Clear': 'מחוק',
+          'Font': 'פונט',
+          'Search': 'חפש',
+          'matches': 'תוצאות',
+          'showing match': 'מציג תוצאות',
+          'Prev': 'הקודם',
+          'Next': 'הבא',
+          'Camera': 'מצלמה',
+          'Video': 'וידאו',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
+        },
+        'ms': {
+          'Paste a link': 'Tampal Pautan',
+          'Ok': 'Ok',
+          'Select Color': 'Pilih Warna',
+          'Gallery': 'Galeri',
+          'Link': 'Pautan',
+          'Please first select some text to transform into a link.':
+              'Sila pilih beberapa patah perkataan'
+                  ' untuk diubah menjadi pautan.',
+          'Open': 'Buka',
+          'Copy': 'Salin',
+          'Remove': 'Buang',
+          'Save': 'Simpan',
+          'Zoom': 'Zum',
+          'Saved': 'Telah Disimpan',
+          'Text': 'Perkataan',
+          'What is entered is not a link': 'Apa yang diisi bukan pautan',
+          'Resize': 'Ubah saiz',
+          'Width': 'Lebar',
+          'Height': 'Tinggi',
+          'Size': 'Saiz',
+          'Small': 'Kecil',
+          'Large': 'Besar',
+          'Huge': 'Amat Besar',
+          'Clear': 'Padam',
+          'Font': 'Fon',
+          'Search': 'Carian',
+          'matches': 'padanan',
+          'showing match': 'menunjukkan padanan',
+          'Prev': 'Sebelum',
+          'Next': 'Seterusnya',
+          'Camera': 'Kamera',
+          'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
+        },
+        'it': {
+          'Paste a link': 'Incolla un collegamento',
+          'Ok': 'Ok',
+          'Select Color': 'Seleziona Colore',
+          'Gallery': 'Galleria',
+          'Link': 'Collegamento',
+          'Please first select some text to transform into a link.':
+              'Per prima cosa seleziona del testo da trasformare in un link.',
+          'Open': 'Apri',
+          'Copy': 'Copia',
+          'Remove': 'Rimuovi',
+          'Save': 'Salva',
+          'Zoom': 'Ingrandisci',
+          'Saved': 'Salvato',
+          'Text': 'Testo',
+          'What is entered is not a link':
+              'Ciò che viene inserito non è un collegamento',
+          'Resize': 'Ridimensiona',
+          'Width': 'Larghezza',
+          'Height': 'Altezza',
+          'Size': 'Dimensione',
+          'Small': 'Piccolo',
+          'Large': 'Largo',
+          'Huge': 'Enorme',
+          'Clear': 'Cancella',
+          'Font': 'Font',
+          'Search': 'Ricerca',
+          'matches': 'corrispondenze',
+          'showing match': 'visualizza corrispondenza',
+          'Prev': 'Prec',
+          'Next': 'Succ',
+          'Camera': 'Camera',
+          'Video': 'Video',
+          'Undo': 'Undo',
+          'Redo': 'Redo',
+          'Font family': 'Font family',
+          'Font size': 'Font size',
+          'Bold': 'Bold',
+          'Subscript': 'Subscript',
+          'Superscript': 'Superscript',
+          'Italic': 'Italic',
+          'Underline': 'Underline',
+          'Strike through': 'Strike through',
+          'Inline code': 'Inline code',
+          'Font color': 'Font color',
+          'Background color': 'Background color',
+          'Clear format': 'Clear format',
+          'Align left': 'Align left',
+          'Align center': 'Align center',
+          'Align right': 'Align right',
+          'Justify win width': 'Justify win width',
+          'Text direction': 'Text direction',
+          'Header style': 'Header style',
+          'Numbered list': 'Numbered list',
+          'Bullet list': 'Bullet list',
+          'Checked list': 'Checked list',
+          'Code block': 'Code block',
+          'Quote': 'Quote',
+          'Increase indent': 'Increase indent',
+          'Decrease indent': 'Decrease indent',
+          'Insert URL': 'Insert URL',
+          'Visit link': 'Visit link',
+          'Enter link': 'Enter link',
+          'Enter media': 'Enter media',
+          'Edit': 'Edit',
+          'Apply': 'Apply',
+        },
+        'bn': {
+          'Paste a link': 'লিঙ্ক পেস্ট করুন',
+          'Ok': 'ওকে',
+          'Select Color': 'কালার সিলেক্ট করুন',
+          'Gallery': 'গ্যালারি',
+          'Link': 'লিঙ্ক',
+          'Please first select some text to transform into a link.':
+              'অনুগ্রহ করে প্রথমে একটি লিঙ্কে রূপান্তরিত করার '
+                  'জন্য কিছু পাঠ্য নির্বাচন করুন।',
+          'Open': 'ওপেন',
+          'Copy': 'কপি',
+          'Remove': 'রিমুভ',
+          'Save': 'সেভ',
+          'Zoom': 'জুম',
+          'Saved': 'সেভড',
+          'Text': 'টেক্সট',
+          'What is entered is not a link': 'যা দেওয়া হয়েছে তা একটি লিঙ্ক নয়',
+          'Resize': 'রিসাইজ',
+          'Width': 'প্রস্থ',
+          'Height': 'দৈর্ঘ্য',
+          'Size': 'সাইজ',
+          'Small': 'ছোট',
+          'Large': 'বড়',
+          'Huge': 'বিশাল',
+          'Clear': 'ক্লিয়ার',
+          'Font': 'ফন্ট',
+          'Search': 'সার্চ',
+          'matches': 'মিলে',
+          'showing match': 'মিল দেখানো হচ্ছে',
+          'Prev': 'পূর্ববর্তী',
+          'Next': 'পরবর্তী',
+          'Camera': 'ক্যামেরা',
+          'Video': 'ভিডিও',
+          'Undo': 'আন্ডু',
+          'Redo': 'রিডু',
+          'Font family': 'ফন্ট ফ্যামিলি',
+          'Font size': 'ফন্ট সাইজ',
+          'Bold': 'বোল্ড',
+          'Subscript': 'সাবস্ক্রিপ্ট',
+          'Superscript': 'সুপারস্ক্রিপ্ট',
+          'Italic': 'ইটালিক',
+          'Underline': 'আন্ডারলাইন',
+          'Strike through': 'স্ট্রাইক থ্রু',
+          'Inline code': 'ইনলাইন কোড',
+          'Font color': 'ফন্ট কালার',
+          'Background color': 'ব্যাকগ্রাউন্ড কালার',
+          'Clear format': 'ক্লিয়ার ফরম্যাট',
+          'Align left': 'বাম সারিবদ্ধ',
+          'Align center': 'কেন্দ্র সারিবদ্ধ',
+          'Align right': 'ডান সারিবদ্ধ',
+          'Justify win width': 'প্রস্থের সাথে সংযত',
+          'Text direction': 'টেক্সট ডিরেকশন',
+          'Header style': 'হেডার স্টাইল',
+          'Numbered list': 'সংখ্যাযুক্ত তালিকা',
+          'Bullet list': 'বুলেট তালিকা',
+          'Checked list': 'চেক করা তালিকা',
+          'Code block': 'কোড ব্লক',
+          'Quote': 'উক্তি',
+          'Increase indent': 'ইন্ডেন্ট বাড়ান',
+          'Decrease indent': 'ইন্ডেন্ট কমান',
+          'Insert URL': 'UR দিন',
+          'Visit link': 'ভিজিট লিঙ্ক',
+          'Enter link': 'লিঙ্ক দিন',
+          'Enter media': 'মিডিয়া দিন',
+          'Edit': 'ইডিট',
+          'Apply': 'এপ্লাই',
+          'Hex': 'হেক্স',
+          'Material': 'ম্যাটারিয়াল',
+          'Color': 'কালার',
         },
       };
 
