@@ -1,11 +1,10 @@
 import 'dart:math';
 
-import 'package:tuple/tuple.dart';
-
 import '../models/documents/nodes/leaf.dart';
+import '../models/structs/offset_value.dart';
 import '../widgets/controller.dart';
 
-Tuple2<int, Embed> getEmbedNode(QuillController controller, int offset) {
+OffsetValue<Embed> getEmbedNode(QuillController controller, int offset) {
   var offset = controller.selection.start;
   var embedNode = controller.queryNode(offset);
   if (embedNode == null || !(embedNode is Embed)) {
@@ -13,7 +12,7 @@ Tuple2<int, Embed> getEmbedNode(QuillController controller, int offset) {
     embedNode = controller.queryNode(offset);
   }
   if (embedNode != null && embedNode is Embed) {
-    return Tuple2(offset, embedNode);
+    return OffsetValue(offset, embedNode);
   }
 
   return throw 'Embed node not found by offset $offset';
