@@ -3,18 +3,19 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/scheduler.dart';
 
-import '../../flutter_quill.dart';
+import '../models/documents/attribute.dart';
+import '../models/documents/nodes/leaf.dart';
 import '../utils/platform.dart';
+import 'editor.dart';
+import 'embeds.dart';
 import 'text_selection.dart';
 
-typedef EmbedsBuilder = Widget Function(
-  BuildContext context,
-  QuillController controller,
-  Embed node,
-  bool readOnly,
-);
+typedef EmbedsBuilder = EmbedBuilder Function(Embed node);
 
 typedef CustomStyleBuilder = TextStyle Function(Attribute attribute);
+
+typedef CustomRecognizerBuilder = GestureRecognizer? Function(
+    Attribute attribute, Leaf leaf);
 
 /// Delegate interface for the [EditorTextSelectionGestureDetectorBuilder].
 ///
@@ -66,7 +67,8 @@ class EditorTextSelectionGestureDetectorBuilder {
   /// Creates a [EditorTextSelectionGestureDetectorBuilder].
   ///
   /// The [delegate] must not be null.
-  EditorTextSelectionGestureDetectorBuilder({required this.delegate});
+  EditorTextSelectionGestureDetectorBuilder(
+      {required this.delegate, this.detectWordBoundary = true});
 
   /// The delegate for this [EditorTextSelectionGestureDetectorBuilder].
   ///
@@ -82,6 +84,8 @@ class EditorTextSelectionGestureDetectorBuilder {
   /// will return true if current [onTapDown] event is triggered by a touch or
   /// a stylus.
   bool shouldShowSelectionToolbar = true;
+
+  bool detectWordBoundary = true;
 
   /// The [State] of the [EditableText] for which the builder will provide a
   /// [EditorTextSelectionGestureDetector].
@@ -308,7 +312,8 @@ class EditorTextSelectionGestureDetectorBuilder {
   ///  which triggers this callback./lib/src/material/text_field.dart
   @protected
   void onDragSelectionUpdate(
-      DragStartDetails startDetails, DragUpdateDetails updateDetails) {
+      //DragStartDetails startDetails,
+      DragUpdateDetails updateDetails) {
     renderEditor!.extendSelection(updateDetails.globalPosition,
         cause: SelectionChangedCause.drag);
   }
@@ -337,24 +342,28 @@ class EditorTextSelectionGestureDetectorBuilder {
   ///
   /// The [child] or its subtree should contain [EditableText].
   Widget build(
-      {required HitTestBehavior behavior, required Widget child, Key? key}) {
+      {required HitTestBehavior behavior,
+      required Widget child,
+      Key? key,
+      bool detectWordBoundary = true}) {
     return EditorTextSelectionGestureDetector(
-      key: key,
-      onTapDown: onTapDown,
-      onForcePressStart: delegate.forcePressEnabled ? onForcePressStart : null,
-      onForcePressEnd: delegate.forcePressEnabled ? onForcePressEnd : null,
-      onSingleTapUp: onSingleTapUp,
-      onSingleTapCancel: onSingleTapCancel,
-      onSingleLongTapStart: onSingleLongTapStart,
-      onSingleLongTapMoveUpdate: onSingleLongTapMoveUpdate,
-      onSingleLongTapEnd: onSingleLongTapEnd,
-      onDoubleTapDown: onDoubleTapDown,
-      onSecondarySingleTapUp: onSecondarySingleTapUp,
-      onDragSelectionStart: onDragSelectionStart,
-      onDragSelectionUpdate: onDragSelectionUpdate,
-      onDragSelectionEnd: onDragSelectionEnd,
-      behavior: behavior,
-      child: child,
-    );
+        key: key,
+        onTapDown: onTapDown,
+        onForcePressStart:
+            delegate.forcePressEnabled ? onForcePressStart : null,
+        onForcePressEnd: delegate.forcePressEnabled ? onForcePressEnd : null,
+        onSingleTapUp: onSingleTapUp,
+        onSingleTapCancel: onSingleTapCancel,
+        onSingleLongTapStart: onSingleLongTapStart,
+        onSingleLongTapMoveUpdate: onSingleLongTapMoveUpdate,
+        onSingleLongTapEnd: onSingleLongTapEnd,
+        onDoubleTapDown: onDoubleTapDown,
+        onSecondarySingleTapUp: onSecondarySingleTapUp,
+        onDragSelectionStart: onDragSelectionStart,
+        onDragSelectionUpdate: onDragSelectionUpdate,
+        onDragSelectionEnd: onDragSelectionEnd,
+        behavior: behavior,
+        detectWordBoundary: detectWordBoundary,
+        child: child);
   }
 }
