@@ -6,16 +6,17 @@ import 'package:flutter/gestures.dart';
 import 'package:flutter/material.dart';
 import 'package:flutter/rendering.dart';
 import 'package:flutter/services.dart';
-import 'package:tuple/tuple.dart';
 import 'package:url_launcher/url_launcher.dart';
 
 import '../models/documents/attribute.dart';
 import '../models/documents/nodes/container.dart' as container_node;
+import '../models/documents/nodes/embeddable.dart';
 import '../models/documents/nodes/leaf.dart';
 import '../models/documents/nodes/leaf.dart' as leaf;
 import '../models/documents/nodes/line.dart';
 import '../models/documents/nodes/node.dart';
 import '../models/documents/style.dart';
+import '../models/structs/vertical_spacing.dart';
 import '../utils/color.dart';
 import '../utils/font.dart';
 import '../utils/platform.dart';
@@ -40,6 +41,8 @@ class TextLine extends StatefulWidget {
     required this.linkActionPicker,
     this.textDirection,
     this.customStyleBuilder,
+    this.customRecognizerBuilder,
+    this.customLinkPrefixes = const <String>[],
     Key? key,
   }) : super(key: key);
 
@@ -50,8 +53,10 @@ class TextLine extends StatefulWidget {
   final bool readOnly;
   final QuillController controller;
   final CustomStyleBuilder? customStyleBuilder;
+  final CustomRecognizerBuilder? customRecognizerBuilder;
   final ValueChanged<String>? onLaunchUrl;
   final LinkActionPicker linkActionPicker;
+  final List<String> customLinkPrefixes;
 
   @override
   State<TextLine> createState() => _TextLineState();
@@ -132,17 +137,23 @@ class _TextLineState extends State<TextLine> {
   @override
   Widget build(BuildContext context) {
     assert(debugCheckHasMediaQuery(context));
+
     if (widget.line.hasEmbed && widget.line.childCount == 1) {
-      // For video, it is always single child
-      final embed = widget.line.children.single as Embed;
-      return EmbedProxy(
-        widget.embedBuilder(
-          context,
-          widget.controller,
-          embed,
-          widget.readOnly,
-        ),
-      );
+      // Single child embeds can be expanded
+      var embed = widget.line.children.single as Embed;
+      // Creates correct node for custom embed
+      if (embed.value.type == BlockEmbed.customType) {
+        embed = Embed(CustomBlockEmbed.fromJsonString(embed.value.data));
+      }
+      final embedBuilder = widget.embedBuilder(embed);
+      if (embedBuilder.expanded) {
+        // Creates correct node for custom embed
+        final lineStyle = _getLineStyle(widget.styles);
+        return EmbedProxy(
+          embedBuilder.build(context, widget.controller, embed, widget.readOnly,
+              false, lineStyle),
+        );
+      }
     }
     final textSpan = _getTextSpanForWholeLine(context);
     final strutStyle = StrutStyle.fromTextStyle(textSpan.style!);
@@ -173,24 +184,30 @@ class _TextLineState extends State<TextLine> {
     // The line could contain more than one Embed & more than one Text
     final textSpanChildren = <InlineSpan>[];
     var textNodes = LinkedList<Node>();
-    for (final child in widget.line.children) {
+    for (var child in widget.line.children) {
       if (child is Embed) {
         if (textNodes.isNotEmpty) {
           textSpanChildren
               .add(_buildTextSpan(widget.styles, textNodes, lineStyle));
           textNodes = LinkedList<Node>();
         }
-        // Here it should be image
-        final embed = WidgetSpan(
-          child: EmbedProxy(
-            widget.embedBuilder(
-              context,
-              widget.controller,
-              child,
-              widget.readOnly,
-            ),
+        // Creates correct node for custom embed
+        if (child.value.type == BlockEmbed.customType) {
+          child = Embed(CustomBlockEmbed.fromJsonString(child.value.data))
+            ..applyStyle(child.style);
+        }
+        final embedBuilder = widget.embedBuilder(child);
+        final embedWidget = EmbedProxy(
+          embedBuilder.build(
+            context,
+            widget.controller,
+            child,
+            widget.readOnly,
+            true,
+            lineStyle,
           ),
         );
+        final embed = embedBuilder.buildWidgetSpan(embedWidget);
         textSpanChildren.add(embed);
         continue;
       }
@@ -262,7 +279,7 @@ class _TextLineState extends State<TextLine> {
       toMerge = defaultStyles.quote!.style;
     } else if (block == Attribute.codeBlock) {
       toMerge = defaultStyles.code!.style;
-    } else if (block == Attribute.list) {
+    } else if (block?.key == Attribute.list.key) {
       toMerge = defaultStyles.lists!.style;
     }
 
@@ -295,12 +312,14 @@ class _TextLineState extends State<TextLine> {
     final isLink = nodeStyle.containsKey(Attribute.link.key) &&
         nodeStyle.attributes[Attribute.link.key]!.value != null;
 
+    final recognizer = _getRecognizer(node, isLink);
+
     return TextSpan(
       text: textNode.value,
       style: _getInlineTextStyle(
           textNode, defaultStyles, nodeStyle, lineStyle, isLink),
-      recognizer: isLink && canLaunchLinks ? _getRecognizer(node) : null,
-      mouseCursor: isLink && canLaunchLinks ? SystemMouseCursors.click : null,
+      recognizer: recognizer,
+      mouseCursor: (recognizer != null) ? SystemMouseCursors.click : null,
     );
   }
 
@@ -333,6 +352,14 @@ class _TextLineState extends State<TextLine> {
         }
       }
     });
+
+    if (nodeStyle.containsKey(Attribute.script.key)) {
+      if (nodeStyle.attributes.values.contains(Attribute.subscript)) {
+        res = _merge(res, defaultStyles.subscript!);
+      } else if (nodeStyle.attributes.values.contains(Attribute.superscript)) {
+        res = _merge(res, defaultStyles.superscript!);
+      }
+    }
 
     if (nodeStyle.containsKey(Attribute.inlineCode.key)) {
       res = _merge(res, defaultStyles.inlineCode!.styleFor(lineStyle));
@@ -380,19 +407,38 @@ class _TextLineState extends State<TextLine> {
     return res;
   }
 
-  GestureRecognizer _getRecognizer(Node segment) {
+  GestureRecognizer? _getRecognizer(Node segment, bool isLink) {
     if (_linkRecognizers.containsKey(segment)) {
       return _linkRecognizers[segment]!;
     }
 
-    if (isDesktop() || widget.readOnly) {
-      _linkRecognizers[segment] = TapGestureRecognizer()
-        ..onTap = () => _tapNodeLink(segment);
-    } else {
-      _linkRecognizers[segment] = LongPressGestureRecognizer()
-        ..onLongPress = () => _longPressLink(segment);
+    if (widget.customRecognizerBuilder != null) {
+      final textNode = segment as leaf.Text;
+      final nodeStyle = textNode.style;
+
+      nodeStyle.attributes.forEach((key, value) {
+        final recognizer = widget.customRecognizerBuilder!.call(value, segment);
+        if (recognizer != null) {
+          _linkRecognizers[segment] = recognizer;
+          return;
+        }
+      });
     }
-    return _linkRecognizers[segment]!;
+
+    if (_linkRecognizers.containsKey(segment)) {
+      return _linkRecognizers[segment]!;
+    }
+
+    if (isLink && canLaunchLinks) {
+      if (isDesktop() || widget.readOnly) {
+        _linkRecognizers[segment] = TapGestureRecognizer()
+          ..onTap = () => _tapNodeLink(segment);
+      } else {
+        _linkRecognizers[segment] = LongPressGestureRecognizer()
+          ..onLongPress = () => _longPressLink(segment);
+      }
+    }
+    return _linkRecognizers[segment];
   }
 
   Future<void> _launchUrl(String url) async {
@@ -414,7 +460,7 @@ class _TextLineState extends State<TextLine> {
     launchUrl ??= _launchUrl;
 
     link = link.trim();
-    if (!linkPrefixes
+    if (!(widget.customLinkPrefixes + linkPrefixes)
         .any((linkPrefix) => link!.toLowerCase().startsWith(linkPrefix))) {
       link = 'https://$link';
     }
@@ -476,7 +522,7 @@ class EditableTextLine extends RenderObjectWidget {
   final Widget? leading;
   final Widget body;
   final double indentWidth;
-  final Tuple2 verticalSpacing;
+  final VerticalSpacing verticalSpacing;
   final TextDirection textDirection;
   final TextSelection textSelection;
   final Color color;
@@ -526,8 +572,8 @@ class EditableTextLine extends RenderObjectWidget {
   EdgeInsetsGeometry _getPadding() {
     return EdgeInsetsDirectional.only(
         start: indentWidth,
-        top: verticalSpacing.item1,
-        bottom: verticalSpacing.item2);
+        top: verticalSpacing.top,
+        bottom: verticalSpacing.bottom);
   }
 }
 
@@ -1033,9 +1079,16 @@ class RenderEditableTextLine extends RenderEditableBox {
   @override
   void paint(PaintingContext context, Offset offset) {
     if (_leading != null) {
-      final parentData = _leading!.parentData as BoxParentData;
-      final effectiveOffset = offset + parentData.offset;
-      context.paintChild(_leading!, effectiveOffset);
+      if (textDirection == TextDirection.ltr) {
+        final parentData = _leading!.parentData as BoxParentData;
+        final effectiveOffset = offset + parentData.offset;
+        context.paintChild(_leading!, effectiveOffset);
+      } else {
+        final parentData = _leading!.parentData as BoxParentData;
+        final effectiveOffset = offset + parentData.offset;
+        context.paintChild(_leading!,
+            Offset(size.width - _leading!.size.width, effectiveOffset.dy));
+      }
     }
 
     if (_body != null) {
@@ -1091,6 +1144,18 @@ class RenderEditableTextLine extends RenderEditableBox {
         _selectedRects ??= _body!.getBoxesForSelection(
           local,
         );
+
+        // Paint a small rect at the start of empty lines that
+        // are contained by the selection.
+        if (line.isEmpty &&
+            textSelection.baseOffset <= line.offset &&
+            textSelection.extentOffset > line.offset) {
+          final lineHeight =
+              preferredLineHeight(TextPosition(offset: line.offset));
+          _selectedRects
+              ?.add(TextBox.fromLTRBD(0, 0, 3, lineHeight, textDirection));
+        }
+
         _paintSelection(context, effectiveOffset);
       }
     }
