@@ -1,5 +1,7 @@
 import 'package:flutter/material.dart';
 
+import '../../utils/widgets.dart';
+
 class QuillIconButton extends StatelessWidget {
   const QuillIconButton({
     required this.onPressed,
@@ -10,6 +12,7 @@ class QuillIconButton extends StatelessWidget {
     this.hoverElevation = 1,
     this.highlightElevation = 1,
     this.borderRadius = 2,
+    this.tooltip,
     Key? key,
   }) : super(key: key);
 
@@ -21,24 +24,28 @@ class QuillIconButton extends StatelessWidget {
   final double hoverElevation;
   final double highlightElevation;
   final double borderRadius;
+  final String? tooltip;
 
   @override
   Widget build(BuildContext context) {
     return ConstrainedBox(
       constraints: BoxConstraints.tightFor(width: size, height: size),
-      child: RawMaterialButton(
-        visualDensity: VisualDensity.compact,
-        shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(borderRadius)),
-        fillColor: fillColor,
-        elevation: 0,
-        hoverElevation: hoverElevation,
-        highlightElevation: hoverElevation,
-        onPressed: () {
-          onPressed?.call();
-          afterPressed?.call();
-        },
-        child: icon,
+      child: UtilityWidgets.maybeTooltip(
+        message: tooltip,
+        child: RawMaterialButton(
+          visualDensity: VisualDensity.compact,
+          shape: RoundedRectangleBorder(
+              borderRadius: BorderRadius.circular(borderRadius)),
+          fillColor: fillColor,
+          elevation: 0,
+          hoverElevation: hoverElevation,
+          highlightElevation: hoverElevation,
+          onPressed: () {
+            onPressed?.call();
+            afterPressed?.call();
+          },
+          child: icon,
+        ),
       ),
     );
   }
